refactor: move app code under src
