+++ src/app/api/chat/[chatId]/route.ts
