+++ src/app/api/chat/create/route.ts
