+++ src/app/api/chat/openrouter/route.ts
