+++ src/components/chat/ConversationContainer.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { ChevronDown, Sparkles, Loader2 } from "lucide-react";
 import Image from "next/image";
-import InputBar from "../components/InputBar";
-import ConversationComp from "../components/ConversationComp";
-import ConversationStarter from "../components/ConversationStarter";
-import siteIcon from "../assets/site-icon.png";
-import sampleData from "../assets/sampleData.json";
-import useConversation, { ModelType } from "../hooks/useConversation";
-import useOpenRouterModels from "../hooks/useOpenRouterModels";
-import useGroqModels from "../hooks/useGroqModels";
+import InputBar from "./InputBar";
+import ConversationComp from "./ConversationComp";
+import ConversationStarter from "./ConversationStarter";
+import siteIcon from "../../assets/site-icon.png";
+import sampleData from "../../assets/sampleData.json";
+import useConversation, { ModelType } from "../../hooks/useConversation";
+import useOpenRouterModels from "../../hooks/useOpenRouterModels";
+import useGroqModels from "../../hooks/useGroqModels";
 import clsx from "clsx";
 import { useRouter } from 'next/navigation';
