+++ tsconfig.json
@@ -37,6 +37,9 @@
       ],
       "@/assets/*": [
         "./src/assets/*"
+      ],
+      "@/models/*": [
+        "./src/server/models/*"
       ]
     },
     "target": "ES2017"
